This is a fork of NCReport version 0.7.30,
I don't know  if is the last GPL version I can't find other source, but I think that is a good library to make nice reports in C++ with Qt Framework.
NociSoft have a version more updated but is comercial. If you want to know how to use or contribute to improve this library, you are welcome to develop.

Regards,